<https://datatracker.ietf.org/doc/html/rfc5321#section-4.3.2>
RCPT
         E: 550, 551, 552, 553, 450, 451, 452, 503, 455, 555

<https://www.iana.org/assignments/smtp-enhanced-status-codes/smtp-enhanced-status-codes.xhtml>

X.7.1 	Delivery not authorized, message refused 	451, 454, 502, 503, 533, 550, 551 	The sender is not authorized to send to the destination. This can be the result of per-host or per-recipient filtering. This memo does not discuss the merits of any such filtering, but provides a mechanism to report such. This is useful only as a permanent error. 	[RFC3463] (Standards Track) 	G. Vaudreuil 	IESG

postfix.org:
RCPT TO:
554 5.7.1 <donaldl@digilicious.com>: Relay access denied

sendmail.org:
RCPT TO:
550 5.7.1 <donaldl@digilicious.com>... Relaying denied

meta1.org / esmtp.org
RCPT TO:
550 5.7.1 Relaying denied.

gmail.com
RCPT TO:
550-5.1.1 The email account that you tried to reach does not exist. Please try
550-5.1.1 double-checking the recipient's email address for typos or
550-5.1.1 unnecessary spaces. Learn more at
550 5.1.1  https://support.google.com/mail/?p=NoSuchUser jx13si2647658qvb.159 - gsmtp

exim.org
RCPT TO:
550 relay not permitted

yahoo.com
RCPT TO:
550 relaying denied for <donaldl@digilicious.com> tnmpmscs

outlook.com
RCPT TO:
550 5.5.0 Requested action not taken: mailbox unavailable (S2017062302). [DM6NAM11FT039.eop-nam11.prod.protection.outlook.com]

fastmail.com / messagingengine.com
RCPT TO:
451 4.7.1 <donaldl@digilicious.com>: Relay access denied