The istring and string PEGTL matchers require of the input stream that it buffer at least the length of the string before any match is attempted. This is true even if the contents of the buffer contain enough data to show the string does NOT match.
So, when matching SMTP commands, we make sure to have at least 4 bytes in the buffer; after that we look for single bytes at a time, to not wait for more data that we may not need.

In Session::cmd_unrecognized() we look to see if the data resembles a command, but maybe failed to match due to bad syntax in a domain name or email address to give a “better” error message.